requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg[binary]>=3.2.9   # latest stable with binary wheels
pydantic==2.8.2
python-dotenv==1.0.1
